requirements.txt
flask==2.0.1               
flask-cors==3.0.10        

pyvis==0.3.1              
networkx==2.8.4            

jinja2==3.0.1         
werkzeug==2.0.1          
click==8.0.1            
itsdangerous==2.0.1       
markupsafe==2.0.1         

six==1.16.0               
python-dotenv==0.19.0      
pytz
requests
bs4
openai